ors/requirement.txt
Babel==2.5.3
certifi==2018.1.18
chardet==3.0.4
Django==2.0.3
django-crispy-forms==1.7.2
django-paypal==0.5.0
django-phonenumber-field==2.0.0
django-twilio==0.9.0
idna==2.6
phonenumberslite==8.9.2
Pillow==5.0.0
pkg-resources==0.0.0
PyJWT==1.6.1
PySocks==1.6.8
pytz==2018.3
reportlab==3.4.0
requests==2.18.4
six==1.11.0
twilio==6.11.0
urllib3==1.22
